made path treaversal cleaer
