+++ server/src/parse-map-packs.ts
@@ -113,7 +113,7 @@ export function processLine(
         );
       }
       map.name = formatColumnName(String(info[1]));
-      const trimmedPath = map_info.replace(/\/map-packs\/.*$/, "");
+      const trimmedPath = path.dirname(path.dirname(map_info));
       map.img = `/${trimmedPath}/MapImages/${String(info[2])}`;
       map.note = String(info[3]);
       break;
